requirements.txt
folium==0.17.0
geopy==2.4.1
requests==2.32.3
python-dotenv==1.0.1

